Add files via upload
I keep getting an error that it could not find the .csv file.  I tried recreating it, copying path, and copying relative path and it would not resolve the error.  I also tried copying the file in different folders, but that didn't work.  I decided to upload what I did have.